Move struct definitions to more appropriate headers
